MB-65755 convert SNAP_STATS_ARR_ITEMS_COUNT to int64 instead of uint64
As the value being written is int64 in the IndexStats map (in
OpenSnapshot). It has to be read as int64 instead of uint64

+++ secondary/indexer/plasma_slice.go
@@ -4194,7 +4194,7 @@ func (s *plasmaSnapshot) CountTotal(ctx IndexReaderContext, stopch StopChannel) 
 		s.slice.meteringStats.recordReadUsageStats(ru)
 	}
 	if s.slice.idxStats.useArrItemsCount {
-		return s.info.IndexStats[SNAP_STATS_ARR_ITEMS_COUNT].(uint64), nil
+		return uint64(s.info.IndexStats[SNAP_STATS_ARR_ITEMS_COUNT].(int64)), nil
 	}
 	return uint64(s.MainSnap.Count()), nil
 }
